// node-project/types/Books.ts
export type Books = {
    ID : number ;
    Name: String ;
    Author : Array<String>[];
    Genre: Array<String>[];
    Year: Number ;
    Publisher: String ;
    Country: String ;
    Pages :Number; 
}

export type LiteBooks = {
    ID : number ;
    Name: String;
    Author : Array<String>[];
    Genre: Array<String>[];

}

export type name ={
    Name: String;
}

export type author = {
    Author: Array<String>[];
}